requirements.txt
Django==4.2.8
django-cors-headers==4.3.1
environs==10.0.0
djangorestframework==3.14.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
django-simple-history==3.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-magic==0.4.27
django-extensions==3.2.3
django-model-utils==4.3.1
ipython==8.12.3
